+++ packages/myst-to-react/src/dropdown.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import type { NodeRenderer } from './types';
+import type { NodeRenderer } from '@myst-theme/providers';
 import { ChevronRightIcon } from '@heroicons/react/24/solid';
 import classNames from 'classnames';