+++ packages/myst-to-react/src/image.tsx
@@ -1,6 +1,7 @@
-import type { Alignment } from '@curvenote/blocks';
 import type { Image as ImageNodeSpec } from 'myst-spec';
-import type { NodeRenderer } from './types';
+import type { NodeRenderer } from '@myst-theme/providers';
+
+type Alignment = 'left' | 'center' | 'right';
 
 type ImageNode = ImageNodeSpec & { height?: string };