+++ packages/myst-to-react/src/index.tsx
@@ -1,5 +1,5 @@
-import type { NodeRenderer } from './types';
-import type { GenericParent } from 'mystjs';
+import type { NodeRenderer } from '@myst-theme/providers';
+import type { GenericParent } from 'myst-common';
 import { mystToReact } from './convertToReact';
 import BASIC_RENDERERS from './basic';
 import ADMONITION_RENDERERS from './admonitions';
@@ -14,18 +14,14 @@ import REACTIVE_RENDERERS from './reactive';
 import IFRAME_RENDERERS from './iframe';
 import IMAGE_RENDERERS from './image';
 import LINK_RENDERERS from './links';
-import OUTPUT_RENDERERS from './output';
 import HEADING_RENDERERS from './heading';
 import CROSS_REFERENCE_RENDERERS from './crossReference';
 import TAB_RENDERERS from './tabs';
-import MYST_RENDERERS from './myst';
-import MERMAID_RENDERERS from './mermaid';
-export { MySTRenderer } from './myst';
 import EXT_RENDERERS from './extensions';
 
-export type { NodeRenderer } from './types';
 export { CopyIcon } from './components/CopyIcon';
 export { CodeBlock } from './code';
+export { Admonition, AdmonitionKind } from './admonitions';
 
 export const DEFAULT_RENDERERS: Record<string, NodeRenderer> = {
   ...BASIC_RENDERERS,
@@ -39,11 +35,8 @@ export const DEFAULT_RENDERERS: Record<string, NodeRenderer> = {
   ...FOOTNOTE_RENDERERS,
   ...ADMONITION_RENDERERS,
   ...REACTIVE_RENDERERS,
-  ...OUTPUT_RENDERERS,
   ...HEADING_RENDERERS,
   ...CROSS_REFERENCE_RENDERERS,
-  ...MYST_RENDERERS,
-  ...MERMAID_RENDERERS,
   ...DROPDOWN_RENDERERS,
   ...CARD_RENDERERS,
   ...GRID_RENDERERS,