+++ packages/myst-to-react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # myst-to-react
 
+## 0.1.20
+
+### Patch Changes
+
+- Updates for repackaging for jupyterlab-myst
+- Updated dependencies
+  - @myst-theme/providers@0.1.20
+
+## 0.1.19
+
+### Patch Changes
+
+- Change components to use @myst-theme
+- ba0424a: Upgrade heroicons to v2
+  - @myst-theme/providers@0.1.19
+
 ## 0.1.18
 
 ### Patch Changes